refactored from individual to combined
